fixed var_00.py and makefile works
