+++ lab_02_cc_taco_lab/CC_TACO_LAB/lab/objectives/objective-a/task_02/run-clean-all.x
